+++ sqlstore/migrations/0001_initial.sql
@@ -16,23 +16,23 @@ create table assets
     quantum        INT,
     source         TEXT,
     erc20_contract TEXT,
-    vega_time      TIMESTAMP WITH TIME ZONE NOT NULL
+    vega_time      TIMESTAMP WITH TIME ZONE NOT NULL REFERENCES blocks (vega_time)
 );
 
 create table parties
 (
     id        BYTEA NOT NULL PRIMARY KEY,
-    vega_time TIMESTAMP WITH TIME ZONE
+    vega_time TIMESTAMP WITH TIME ZONE REFERENCES blocks (vega_time)
 );
 
 create table accounts
 (
     id        SERIAL PRIMARY KEY,
     party_id  BYTEA,
-    asset_id  BYTEA                    NOT NULL,
+    asset_id  BYTEA                    NOT NULL REFERENCES assets (id),
     market_id BYTEA,
     type      INT,
-    vega_time TIMESTAMP WITH TIME ZONE NOT NULL,
+    vega_time TIMESTAMP WITH TIME ZONE NOT NULL REFERENCES blocks(vega_time),
 
     UNIQUE(party_id, asset_id, market_id, type)
 );
@@ -40,10 +40,10 @@ create table accounts
 create table ledger
 (
     id              SERIAL                   PRIMARY KEY,
-    account_from_id INT                      NOT NULL,
-    account_to_id   INT                      NOT NULL,
+    account_from_id INT                      NOT NULL REFERENCES accounts(id),
+    account_to_id   INT                      NOT NULL REFERENCES accounts(id),
     quantity        NUMERIC(32, 0)           NOT NULL,
-    vega_time       TIMESTAMP WITH TIME ZONE NOT NULL,
+    vega_time       TIMESTAMP WITH TIME ZONE NOT NULL REFERENCES blocks(vega_time),
     transfer_time   TIMESTAMP WITH TIME ZONE NOT NULL,
     reference       TEXT,
     type            TEXT